Fix clippy lints failing CI (unnecessary_cast, ptr_arg, field_reassign_with_default)
- gea-inference bench.rs: drop redundant u64 -> u64 cast on total_us
- gea-cli main.rs: take &Path instead of &PathBuf in run_demo/resolve_path
- gea-cli main.rs: build GeneratorConfig with struct init instead of
  field reassignment after Default::default()

+++ crates/gea-inference/src/bench.rs
@@ -90,7 +90,7 @@ fn summarize(precision: Precision, mut latencies: Vec<u64>, model_bytes: usize) 
     BenchResult {
         precision,
         iterations,
-        total_us: (total_ns / 1000) as u64,
+        total_us: total_ns / 1000,
         mean_us,
         p50_us: p50_us.max(1),
         p99_us: p99_us.max(1),
